feat: dedicated folder to cache types
